accounts/views.py
from rest_framework import viewsets

from accounts import models, serializers


class CustomUserViewSet(viewsets.ModelViewSet):
    queryset = models.CustomUser.objects.all()
    serializer_class = serializers.CustomUserSerializer
